extract bbox calculation into layout
